+++ tests/Unit/Anonymization/Configuration/AttributeLoaderTest.php
@@ -4,43 +4,31 @@ declare(strict_types=1);
 
 namespace MakinaCorpus\DbToolsBundle\Tests\Unit\Anonymization\Configuration;
 
+use Doctrine\Common\EventManager;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadataFactory;
 use Doctrine\ORM\Mapping\Driver\AttributeDriver;
 use Doctrine\ORM\Tools\Console\EntityManagerProvider;
-use Doctrine\Persistence\Mapping\RuntimeReflectionService;
 use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizationConfig;
 use MakinaCorpus\DbToolsBundle\Anonymization\Config\AnonymizerConfig;
 use MakinaCorpus\DbToolsBundle\Anonymization\Config\Loader\AttributesLoader;
 use MakinaCorpus\DbToolsBundle\Tests\Resources\Loader\TestEntity;
 use MakinaCorpus\DbToolsBundle\Test\UnitTestCase;
+use MakinaCorpus\DbToolsBundle\Tests\Resources\Loader\TestEmbeddableEntity;
 use MakinaCorpus\DbToolsBundle\Tests\Resources\Loader\TestEntityWithEmbedded;
 
 class AttributeLoaderTest extends UnitTestCase
 {
     public function testLoadOk(): void
     {
-        $attributeDriver = new AttributeDriver([
-            \dirname(\dirname(\dirname(__DIR__))) . '/Resources/Loader'
-        ]);
-        $classMetadata = new ClassMetadata(TestEntity::class);
-        $classMetadata->initializeReflection(new RuntimeReflectionService());
-
-        $attributeDriver->loadMetadataForClass(TestEntity::class, $classMetadata);
-
-        $metaDataFactory = $this->createMock(ClassMetadataFactory::class);
-        $metaDataFactory
-            ->expects($this->exactly(1))
-            ->method('getAllMetadata')
-            ->willReturn([$classMetadata])
-        ;
+        $classMetadataFactory = $this->getClassMetadataFactory();
+        $classMetadataFactory->getMetadataFor(TestEntity::class);
 
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager
             ->expects($this->exactly(1))
             ->method('getMetadataFactory')
-            ->willReturn($metaDataFactory)
+            ->willReturn($classMetadataFactory)
         ;
 
         $entityManagerProvider = $this->createMock(EntityManagerProvider::class);
@@ -55,8 +43,6 @@ class AttributeLoaderTest extends UnitTestCase
         (new AttributesLoader($entityManagerProvider))->load($config);
 
         // Then we validate what's in it:
-        self::assertCount(1, $config->all());
-
         $testTableConfig = $config->getTableConfig('test');
         self::assertCount(3, $testTableConfig);
 
@@ -82,28 +68,16 @@ class AttributeLoaderTest extends UnitTestCase
         self::assertSame('country', $testTableConfig['address_0']->options->get('country'));
     }
 
-    public function testLoadIgnoreEmbedded(): void
+    public function testLoadWithEmbeddedOk(): void
     {
-        $attributeDriver = new AttributeDriver([
-            \dirname(\dirname(\dirname(__DIR__))) . '/Resources/Loader'
-        ]);
-        $classMetadata = new ClassMetadata(TestEntityWithEmbedded::class);
-        $classMetadata->initializeReflection(new RuntimeReflectionService());
-
-        $attributeDriver->loadMetadataForClass(TestEntity::class, $classMetadata);
-
-        $metaDataFactory = $this->createMock(ClassMetadataFactory::class);
-        $metaDataFactory
-            ->expects($this->exactly(1))
-            ->method('getAllMetadata')
-            ->willReturn([$classMetadata])
-        ;
+        $classMetadataFactory = $this->getClassMetadataFactory();
+        $classMetadataFactory->getMetadataFor(TestEntityWithEmbedded::class);
 
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager
             ->expects($this->exactly(1))
             ->method('getMetadataFactory')
-            ->willReturn($metaDataFactory)
+            ->willReturn($classMetadataFactory)
         ;
 
         $entityManagerProvider = $this->createMock(EntityManagerProvider::class);
@@ -117,8 +91,50 @@ class AttributeLoaderTest extends UnitTestCase
         $config = new AnonymizationConfig('default');
         (new AttributesLoader($entityManagerProvider))->load($config);
 
-        // Then we validate that there is still nothing in the config:
-        // (it means that the embeddable has been correctly ignored)
-        self::assertCount(0, $config->all());
+        // Then we validate what's in it:
+        $testTableConfig = $config->getTableConfig('test_with_embedded');
+        self::assertCount(3, $testTableConfig);
+
+        self::assertInstanceOf(AnonymizerConfig::class, $testTableConfig['email']);
+        self::assertSame('email', $testTableConfig['email']->anonymizer);
+        self::assertSame('email', $testTableConfig['email']->targetName);
+        self::assertSame('toto.com', $testTableConfig['email']->options->get('domain'));
+
+        self::assertInstanceOf(AnonymizerConfig::class, $testTableConfig['embeddableEntity_age']);
+        self::assertSame('integer', $testTableConfig['embeddableEntity_age']->anonymizer);
+        self::assertSame('embeddableEntity_age', $testTableConfig['embeddableEntity_age']->targetName);
+        self::assertSame(0, $testTableConfig['embeddableEntity_age']->options->get('min'));
+        self::assertSame(65, $testTableConfig['embeddableEntity_age']->options->get('max'));
+
+        self::assertInstanceOf(AnonymizerConfig::class, $testTableConfig['embeddableEntity_size']);
+        self::assertSame('integer', $testTableConfig['embeddableEntity_size']->anonymizer);
+        self::assertSame('embeddableEntity_size', $testTableConfig['embeddableEntity_size']->targetName);
+        self::assertSame(60, $testTableConfig['embeddableEntity_size']->options->get('min'));
+        self::assertSame(250, $testTableConfig['embeddableEntity_size']->options->get('max'));
+    }
+
+    private function getClassMetadataFactory(): ClassMetadataFactory
+    {
+        $attributeDriver = new AttributeDriver([
+            \dirname(\dirname(\dirname(__DIR__))) . '/Resources/Loader'
+        ]);
+
+        $configuration = new \Doctrine\ORM\Configuration();
+        $configuration->setMetadataDriverImpl($attributeDriver);
+
+        $entityManager = $this->createMock(EntityManagerInterface::class);
+        $entityManager
+            ->method('getConfiguration')
+            ->willReturn($configuration)
+        ;
+        $entityManager
+            ->method('getEventManager')
+            ->willReturn(new EventManager())
+        ;
+
+        $classMetadataFactory = new ClassMetadataFactory();
+        $classMetadataFactory->setEntityManager($entityManager);
+
+        return $classMetadataFactory;
     }
 }